Add Player and Battle UI improvements
The modifications in this commit include addition of a new `ProgressiveText` class to render multiline text in a box. The `Player` class received a default name and stats. The `Battle` game mode was improved to include a new `BattleBox` UI element and a progressive text that now appears in the BattleBox. The `PlayerStats` UI was refactored to use smaller fonts. A few new classes were added to handle items and menus but are not used at the moment.